add a navigator override field to the bogie

+++ src/main/java/com/crystaelix/simurail/content/bogey/PhysicsBogeyControlOverrides.java
@@ -7,9 +7,9 @@ public class PhysicsBogeyControlOverrides {
 	public boolean overrideBrakeStrength;
 	public boolean overrideSteerValue;
 	public boolean overrideStressMultiplier;
-	private double brakeStrength;
-	private double steerValue;
-	private double stressMultiplier;
+	private float brakeStrength;
+	private float steerValue;
+	private float stressMultiplier;
 
 	public boolean hasOverrides() {
 		return overrideBrakeStrength || overrideSteerValue || overrideStressMultiplier;
@@ -21,11 +21,11 @@ public class PhysicsBogeyControlOverrides {
 		resetStressMultiplier();
 	}
 
-	public double getBrakeStrength() {
+	public float getBrakeStrength() {
 		return brakeStrength;
 	}
 
-	public void setBrakeStrength(double brakeStrength) {
+	public void setBrakeStrength(float brakeStrength) {
 		this.brakeStrength = Math.clamp(brakeStrength, 0, 1);
 		overrideBrakeStrength = true;
 	}
@@ -35,11 +35,11 @@ public class PhysicsBogeyControlOverrides {
 		overrideBrakeStrength = false;
 	}
 
-	public double getSteerValue() {
+	public float getSteerValue() {
 		return steerValue;
 	}
 
-	public void setSteerValue(double steerValue) {
+	public void setSteerValue(float steerValue) {
 		this.steerValue = Math.clamp(steerValue, -1, 1);
 		overrideSteerValue = true;
 	}
@@ -49,11 +49,11 @@ public class PhysicsBogeyControlOverrides {
 		overrideSteerValue = false;
 	}
 
-	public double getStressMultiplier() {
+	public float getStressMultiplier() {
 		return stressMultiplier;
 	}
 
-	public void setStressMultiplier(double stressMultiplier) {
+	public void setStressMultiplier(float stressMultiplier) {
 		this.stressMultiplier = Math.clamp(stressMultiplier, 0, 1);
 		overrideStressMultiplier = true;
 	}
@@ -66,13 +66,13 @@ public class PhysicsBogeyControlOverrides {
 	protected CompoundTag write() {
 		CompoundTag tag = new CompoundTag();
 		if(overrideBrakeStrength) {
-			tag.putDouble("brake_strength", brakeStrength);
+			tag.putFloat("brake_strength", brakeStrength);
 		}
 		if(overrideSteerValue) {
-			tag.putDouble("steer_value", steerValue);
+			tag.putFloat("steer_value", steerValue);
 		}
 		if(overrideStressMultiplier) {
-			tag.putDouble("stress_multiplier", stressMultiplier);
+			tag.putFloat("stress_multiplier", stressMultiplier);
 		}
 		return tag;
 	}
@@ -81,15 +81,15 @@ public class PhysicsBogeyControlOverrides {
 		reset();
 		if(tag.contains("brake_strength")) {
 			overrideBrakeStrength = true;
-			brakeStrength = tag.getDouble("brake_strength");
+			brakeStrength = tag.getFloat("brake_strength");
 		}
 		if(tag.contains("steer_value")) {
 			overrideSteerValue = true;
-			steerValue = tag.getDouble("steer_value");
+			steerValue = tag.getFloat("steer_value");
 		}
 		if(tag.contains("stress_multiplier")) {
 			overrideStressMultiplier = true;
-			stressMultiplier = tag.getDouble("stress_multiplier");
+			stressMultiplier = tag.getFloat("stress_multiplier");
 		}
 	}
 }
